[refactor/#224] Users.ts 파일명 camelCase로 변경 및 import 경로 수정
