+++ src/services/users.ts
@@ -1,9 +1,9 @@
+import api from "./axios";
 import {
   ProfileImageUrlResponse,
   UpdateUserProfileRequest,
   UserProfileResponse,
-} from "@/src/types/Users";
-import api from "./axios";
+} from "@/src/types/users";
 
 /**
  * 현재 로그인한 사용자의 프로필 정보를 가져옵니다.
@@ -42,10 +42,9 @@ export const updateUserProfile = async (
   return response.data;
 };
 
-
 /**
  * 사용자의 프로필 이미지를 업로드합니다.
- * 
+ *
  * @async
  * @param {File} file - 업로드할 이미지 파일
  * @returns {Promise<ProfileImageUrlResponse>} 업로드된 이미지의 URL을 포함한 응답
